Separated annotation types into their own files. Should have done this from the start
