+++ app/models/annotation.js
@@ -6,6 +6,9 @@ require('./publication');
 require('./annotation_status');
 require('./user');
 require('./keyword');
+require('./gene_term_annotation');
+require('./gene_gene_annotation');
+require('./comment_annotation');
 
 let Annotation = bookshelf.model('Annotation', {
 	tableName: 'annotation',
@@ -18,45 +21,10 @@ let Annotation = bookshelf.model('Annotation', {
 	user: function() {
 		return this.hasOne('User', 'submitter_id');
 	},
-	fullAnnotation: function() {
-		return this.morphTo('annotation_id', GeneTermAnnotation, GeneGeneAnnotation, CommentAnnotation);
-	}
-});
-
-let GeneTermAnnotation = bookshelf.model('GeneTermAnnotation', {
-	tableName: 'gene_term_annotation',
-	method: function() {
-		return this.hasOne('Keyword', 'method_id');
-	},
-	keyword: function() {
-		return this.hasOne('Keyword', 'keyword_id');
-	},
-	fullAnnotation: function() {
-		return this.morphOne(Annotation, 'annotation_id');
-	}
-});
-
-let GeneGeneAnnotation = bookshelf.model('GeneGeneAnnotation', {
-	tableName: 'gene_gene_annotation',
-	method: function() {
-		return this.hasOne('Keyword', 'method_id');
-	},
-	fullAnnotation: function() {
-		return this.morphOne(Annotation, 'annotation_id');
-	}
-});
-
-let CommentAnnotation = bookshelf.model('CommentAnnotation', {
-	tableName: 'comment_annotation',
-	fullAnnotation: function() {
-		return this.morphOne(Annotation, 'annotation_id');
+	childData: function() {
+		return this.morphTo('annotation_id', 'GeneTermAnnotation', 'GeneGeneAnnotation', 'CommentAnnotation');
 	}
 });
 
 
-module.exports = {
-	Annotation,
-	GeneTermAnnotation,
-	GeneGeneAnnotation,
-	CommentAnnotation
-};
+module.exports = Annotation;
