+++ app/models/comment_annotation.js
@@ -1,0 +1,15 @@
+"use strict";
+
+const bookshelf = require('../lib/bookshelf');
+
+require('./annotation');
+
+let CommentAnnotation = bookshelf.model('CommentAnnotation', {
+	tableName: 'comment_annotation',
+	parentData: function() {
+		return this.morphOne('Annotation', 'annotation_id');
+	}
+});
+
+
+module.exports = CommentAnnotation;
